requirements.txt
# Data handling
pandas
numpy

# Database
sqlalchemy
psycopg2-binary

# Orchestration
apache-airflow

# Optional: for reading CSV/Excel from multiple formats
openpyxl
